Move the honeypot outside of fields
If you manually added the fields with the @formField directive instead of looping through them with @formFields, you ended up without the honeypot field.

This change moves the honeypot to its own property outside of the fields collection.

The user now has more control over the honeypot field. If they want to use the feature it, they need to add it in the view with @formHoneypot. But things will also continue working without the honeypot field.