Make field properties dynamic
Instead of accessing and setting properties on a static config array, we are now calling the property methods on demand. This is helpful if you change a field in the `hydratedFields` callback. For instance, if you change the options of a select field, the default value will now automatically take those new options into account. Before this change, you had to manually change the default value to reflect the changes made to the options.