Fix flash of success message
The field values are reset when the form is submitted. This leads to the fields being processed by Alpine again, because we are using `x-effect`. This in turn leads to the `update-step-visibility` event being dispatched. And this will effectively reset the success flash message. Which is why we were seing a quick flash. This is only an issue with the wizard form, as we don’t dispatch that event when using the section form.